renamed kotlin base scripts
