+++ app/src/main/AndroidManifest.xml
@@ -40,7 +40,7 @@
       tools:ignore="GoogleAppIndexingWarning">
 
     <activity
-        android:name="com.google.ar.core.examples.kotlin.helloar.HelloArActivity"
+        android:name="com.google.ar.core.examples.kotlin.helloar.MainActivity"
         android:configChanges="orientation|screenSize"
         android:exported="true"
         android:theme="@style/Theme.AppCompat.NoActionBar"
